Disables parallel antenna diagram calculation
There was a bug in which  eventually the antenna diagram would have
sectors in different scales, and the sectors size were exactly the size
that each thread should treat. Probably a bug with emscripten and
barriers.
Weird thing, is that the bug persisted long after I removed this code.
And the sector size would still be proportional to the number of
threads.
Now the bug is gone misteriously. Probably a caching issue with firefox?